+++ src/scene_machine.py
@@ -5,7 +5,6 @@ from launch_controller import LaunchController
 from scenes_base import Scene
 from settings_controller import SettingsController
 from events import EventListener, Event, NewSceneEvent
-from pygame import Surface
 from world import World
 import constants
 import pygame
@@ -20,7 +19,7 @@ class SceneMachine(EventListener):
         self._screen: pygame.Surface = pygame.display.set_mode(
             constants.SCREEN_SIZE)
 
-        self._controller = LaunchController(self._screen, self._world)
+        self._controller = LaunchController(self._screen, self._start_scene())
         self._prev_controller: Controller = None
 
     def notify(self, event: Event) -> None:
@@ -29,13 +28,23 @@ class SceneMachine(EventListener):
         elif isinstance(event, NewSceneEvent):
             self._set_next_scene(event.scene)
 
+    def _start_scene(self) -> DecisionScene:
+        options = {}
+        for option_key in range(4):
+            scene_name = str(option_key)
+            options[scene_name] = DecisionOption(self._world.scene_count)
+            self._world.scene_count += 1
+
+        main_text = (
+            'scene {}: this is a very long description of an a scene and it '
+            'includes a newline.\nwhat a compelling decision i must '
+            'make.'.format(self._world.current_scene))
+        return DecisionScene(main_text, options)
+
     def _set_next_scene(self, scene: Scene) -> None:
         assert isinstance(scene, DecisionScene)
         self._controller = DecisionSceneController(self._screen, self._world,
                                                    scene)
-
-        # self._controller = self._build_scene(
-        #     self._world, self._screen)
 
     def _toggle_settings(self) -> None:
         if self._prev_controller is None:
@@ -46,27 +55,3 @@ class SceneMachine(EventListener):
 
         self._controller.activate()
         self._prev_controller.deactivate()
-
-    def _build_scene(
-            self,
-            world: World,
-            screen: Surface) -> DecisionSceneController:
-        # Builds 3 options for keys 0,1 and 2 that just point to a new scene
-        # number. If the option is selected it modifies the world object.
-        # The decision_scene_controller listens for key presses and if the name
-        # of the key pressed matches the option_key then it executes the
-        # corresponding action. Feel free to change any of this.
-
-        options = {}
-        for option_key in range(4):
-            scene_name = str(option_key)
-            options[scene_name] = DecisionOption(world.scene_count)
-            world.scene_count += 1
-
-        main_text = (
-            'scene {}: this is a very long description of an a scene and it '
-            'includes a newline.\nwhat a compelling decision i must '
-            'make.'.format(world.current_scene))
-        scene = DecisionScene(main_text, options)
-
-        return DecisionSceneController(screen, world, scene)
